+++ api/src/main/java/org/openmrs/module/mdrtb/form/custom/TransferOutForm.java
@@ -49,7 +49,7 @@ public class TransferOutForm extends AbstractSimpleForm {
 		}
 		
 		// we only need to update this if this is a new obs or if the value has changed.
-		if (obs == null || obs.getValueNumeric() == null || obs.getValueNumeric().intValue() != id.intValue()) {
+		if (obs == null || obs.getValueNumeric() == null || obs.getValueNumeric().intValue() != id) {
 			
 			// void the existing obs if it exists
 			// (we have to do this manually because openmrs doesn't void obs when saved via encounters)
